Split off the random access iterable to separate module.
